requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
chromadb==0.4.24
streamlit==1.35.0
openai==1.30.0
python-dotenv==1.0.1
pypdf==4.2.0
tiktoken==0.7.0